+++ messages/ru-RU/app/modules/blog.php
@@ -36,6 +36,7 @@ return [
     'Updating post: {name}' => "Редактирование публикации: {name}",
     'View blog post' => "Просмотр публикации",
 
+    'Posts' => "Публикации",
     'Categories' => "Рубрики",
     'Categories of posts' => "Рубрики записей",
     'Tags' => "Теги",
@@ -108,5 +109,14 @@ return [
     'No posts available' => "Нет доступных для отображения публикаций",
     'An error occurred while retrieving the posts list' => "Произошла ошибка при получении списка публикаций",
 
+    'Common' => "Дополнительно",
+    'Present in sitemap' => "Присутствует в карте сайта",
+    'Not present in sitemap' => "Не присутствует в карте сайта",
+    'Present in RSS-feed' => "Присутствует в RSS-ленте",
+    'Not present in RSS-feed' => "Не присутствует в RSS-ленте",
+    'Present in Yandex.Turbo' => "Присутствует в Яндекс.Турбо",
+    'Not present in Yandex.Turbo' => "Не присутствует в Яндекс.Турбо",
+    'Present in Google AMP' => "Присутствует в Google AMP",
+    'Not present in Google AMP' => "Не присутствует в Google AMP",
 
 ];