+++ views/posts/_form.php
@@ -3,7 +3,7 @@
 use wdmg\widgets\TagsInput;
 use yii\helpers\Html;
 use yii\helpers\Url;
-use yii\bootstrap\ActiveForm;
+use yii\widgets\ActiveForm;
 use wdmg\widgets\Editor;
 use wdmg\widgets\SelectInput;
 
@@ -11,23 +11,23 @@ use wdmg\widgets\SelectInput;
 /* @var $model wdmg\blog\models\Blog */
 /* @var $form yii\widgets\ActiveForm */
 ?>
-
-    <div class="blog-form">
-        <?php $form = ActiveForm::begin([
-            'id' => "addPostForm",
-            'enableAjaxValidation' => true,
-            'options' => [
-                'enctype' => 'multipart/form-data'
-            ]
-        ]); ?>
+<div class="blog-form row">
+    <?php $form = ActiveForm::begin([
+        'id' => "addPostForm",
+        'enableAjaxValidation' => true,
+        'options' => [
+            'enctype' => 'multipart/form-data'
+        ]
+    ]); ?>
+    <div class="col-xs-12 col-sm-12 col-md-8 col-lg-9">
         <?= $form->field($model, 'name')->textInput(['maxlength' => true]) ?>
         <?php
             $output = '';
             if (($postURL = $model->getPostUrl(true, true)) && $model->id) {
                 $output = Html::a($model->getPostUrl(true, false), $postURL, [
-                        'target' => '_blank',
-                        'data-pjax' => 0
-                    ]);
+                    'target' => '_blank',
+                    'data-pjax' => 0
+                ]);
             }
 
             if (!empty($output))
@@ -52,7 +52,24 @@ use wdmg\widgets\SelectInput;
             }
         ?>
 
-        <?= $form->field($model, 'categories')->inline(false)->checkboxList($categoriesList) ?>
+        <?= $form->field($model, 'title')->textInput() ?>
+        <?= $form->field($model, 'description')->textarea(['rows' => 3]) ?>
+        <?= $form->field($model, 'keywords')->textarea(['rows' => 3]) ?>
+        <hr/>
+        <div class="form-group">
+            <?= Html::a(Yii::t('app/modules/blog', '&larr; Back to list'), ['posts/index'], ['class' => 'btn btn-default pull-left']) ?>&nbsp;
+            <?= Html::submitButton(Yii::t('app/modules/blog', 'Save'), ['class' => 'btn btn-success pull-right']) ?>
+        </div>
+    </div>
+    <div class="col-xs-12 col-sm-12 col-md-4 col-lg-3">
+
+        <?= $form->field($model, 'categories')->checkboxList($categoriesList, [
+            'class' => 'list-group',
+            'onclick' => "$(this).val( $('input:checkbox:checked').val()); ", // if you use required as a validation rule you will need this for the time being until a fix is in place by yii2
+            'item' => function($index, $label, $name, $checked, $value) {
+                return '<li class="list-group-item"><label><input type="checkbox" ' . (($checked) ? "checked" : "") . ' name="' . $name . '" value="' . $value . '" tabindex="' . $index . '">&nbsp;' . $label . '</label></li>';
+            }
+        ]) ?>
 
         <?= $form->field($model, 'tags')->widget(TagsInput::class, [
             'options' => [
@@ -66,18 +83,9 @@ use wdmg\widgets\SelectInput;
             ]
         ])->input('text', ['placeholder' => Yii::t('app/modules/blog', 'Type tags...')]); ?>
 
-        <?php // ])->input('text', ['value' => \yii\helpers\Json::encode($tagsList), 'placeholder' => Yii::t('app/modules/blog', 'Type tags...')]); ?>
-
-
-        <?= $form->field($model, 'title')->textInput() ?>
-        <?= $form->field($model, 'description')->textarea(['rows' => 3]) ?>
-        <?= $form->field($model, 'keywords')->textarea(['rows' => 3]) ?>
-
         <?= $form->field($model, 'in_sitemap', [
-                'template' => "{label}\n<br/>{input}\n{hint}\n{error}"
-            ])
-            ->checkbox(['label' => Yii::t('app/modules/blog', '- display in the sitemap')])
-            ->label(Yii::t('app/modules/blog', 'Sitemap'))
+            'template' => "{label}\n<br/>{input}\n{hint}\n{error}"
+        ])->checkbox(['label' => Yii::t('app/modules/blog', '- display in the sitemap')])->label(Yii::t('app/modules/blog', 'Sitemap'))
         ?>
         <?= $form->field($model, 'in_rss', [
             'template' => "{label}\n<br/>{input}\n{hint}\n{error}"
@@ -100,11 +108,11 @@ use wdmg\widgets\SelectInput;
         ]); ?>
         <hr/>
         <div class="form-group">
-            <?= Html::a(Yii::t('app/modules/blog', '&larr; Back to list'), ['posts/index'], ['class' => 'btn btn-default pull-left']) ?>&nbsp;
-            <?= Html::submitButton(Yii::t('app/modules/blog', 'Save'), ['class' => 'btn btn-success pull-right']) ?>
+            <?= Html::submitButton(Yii::t('app/modules/blog', 'Save'), ['class' => 'btn btn-block btn-success pull-right']) ?>
         </div>
-        <?php ActiveForm::end(); ?>
     </div>
+    <?php ActiveForm::end(); ?>
+</div>
 
 <?php $this->registerJs(<<< JS
 $(document).ready(function() {