+++ views/posts/view.php
@@ -61,10 +61,10 @@ $this->params['breadcrumbs'][] = $this->title;
                 'label' => Yii::t('app/modules/blog', 'Categories'),
                 'format' => 'html',
                 'value' => function($data) {
-                    if ($categories = $data->categories) {
+                    if ($categories = $data->getCategories()) {
                         $output = [];
                         foreach ($categories as $category) {
-                            $output[] = Html::a($category->name, $category->url);
+                            $output[] = Html::a($category->name, ['cats/view', 'id' => $category->id]);
                         }
                         return implode(", ", $output);
                     } else {
@@ -77,10 +77,10 @@ $this->params['breadcrumbs'][] = $this->title;
                 'label' => Yii::t('app/modules/blog', 'Tags'),
                 'format' => 'html',
                 'value' => function($data) {
-                    if ($tags = $data->tags) {
+                    if ($tags = $data->getTags()) {
                         $output = [];
                         foreach ($tags as $tag) {
-                            $output[] = Html::a($tag->name, $tag->url);
+                            $output[] = Html::a($tag->name, ['tags/view', 'id' => $tag->id]);
                         }
                         return implode(", ", $output);
                     } else {
@@ -91,46 +91,43 @@ $this->params['breadcrumbs'][] = $this->title;
 
             'description:ntext',
             'keywords:ntext',
+
             [
-                'attribute' => 'in_sitemap',
+                'attribute' => 'common',
+                'label' => Yii::t('app/modules/blog','Common'),
                 'format' => 'html',
                 'value' => function($data) {
+                    $output = '';
                     if ($data->in_sitemap)
-                        return '<span class="fa fa-check text-success"></span>';
+                        $output .= '<span class="fa fa-fw fa-sitemap text-success" title="' . Yii::t('app/modules/blog','Present in sitemap') . '"></span>';
                     else
-                        return '<span class="fa fa-times text-danger"></span>';
-                }
-            ],
-            [
-                'attribute' => 'in_rss',
-                'format' => 'html',
-                'value' => function($data) {
+                        $output .= '<span class="fa fa-fw fa-sitemap text-danger" title="' . Yii::t('app/modules/blog','Not present in sitemap') . '"></span>';
+
+                    $output .= "&nbsp;";
+
                     if ($data->in_rss)
-                        return '<span class="fa fa-check text-success"></span>';
+                        $output .= '<span class="fa fa-fw fa-rss text-success" title="' . Yii::t('app/modules/blog','Present in RSS-feed') . '"></span>';
                     else
-                        return '<span class="fa fa-times text-danger"></span>';
-                }
-            ],
-            [
-                'attribute' => 'in_turbo',
-                'format' => 'html',
-                'value' => function($data) {
+                        $output .= '<span class="fa fa-fw fa-rss text-danger" title="' . Yii::t('app/modules/blog','Not present in RSS-feed') . '"></span>';
+
+                    $output .= "&nbsp;";
+
                     if ($data->in_turbo)
-                        return '<span class="fa fa-check text-success"></span>';
+                        $output .= '<span class="fa fa-fw fa-rocket text-success" title="' . Yii::t('app/modules/blog','Present in Yandex.Turbo') . '"></span>';
                     else
-                        return '<span class="fa fa-times text-danger"></span>';
-                }
-            ],
-            [
-                'attribute' => 'in_amp',
-                'format' => 'html',
-                'value' => function($data) {
+                        $output .= '<span class="fa fa-fw fa-rocket text-danger" title="' . Yii::t('app/modules/blog','Not present in Yandex.Turbo') . '"></span>';
+
+                    $output .= "&nbsp;";
+
                     if ($data->in_amp)
-                        return '<span class="fa fa-check text-success"></span>';
+                        $output .= '<span class="fa fa-fw fa-bolt text-success" title="' . Yii::t('app/modules/blog','Present in Google AMP') . '"></span>';
                     else
-                        return '<span class="fa fa-times text-danger"></span>';
+                        $output .= '<span class="fa fa-fw fa-bolt text-danger" title="' . Yii::t('app/modules/blog','Not present in Google AMP') . '"></span>';
+
+                    return $output;
                 }
             ],
+
             [
                 'attribute' => 'status',
                 'format' => 'html',